requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django~=3.2.0
django-tinymce==3.4.0
mimesis==5.3.0
Pillow==9.0.1
psycopg2==2.9.3
sqlparse==0.4.2
tzdata==2021.5
